working on avoiding decimal data types

+++ src/cpp/services/dataRow.h
@@ -12,14 +12,14 @@ struct DataRow {
     OrderType orderType;
     std::time_t tradeDate;
     std::time_t settlementDate;
-    double price;
+    long long price;
     int quantity;
-    double fee;
-    double profit;
+    long long fee;
+    long long profit;
     int seq;
     int tempQuantity;
-    double tempFee;
-    double cgt;
+    long long tempFee;
+    long long cgt;
     long long consideration;
     bool twelveMonths = false;
     
@@ -29,7 +29,6 @@ struct DataRow {
     static bool descending(const DataRow& obj, const DataRow& obj2);
     static std::string orderTypeToString(OrderType type);
     static std::string dateToString(std::time_t date);
-
     NLOHMANN_DEFINE_TYPE_INTRUSIVE(DataRow, ASXCode, orderType, tradeDate, settlementDate, price, quantity, fee, profit, seq, tempQuantity, tempFee, cgt, consideration, twelveMonths)
 };
